Add support for the Get_Report Request
The support of the Get_Report request that can be received via the Control pipe is mandatory in the HID specifications. However, it seems that the Linux kernel does not use this request if the device has an IN endpoint. To respect the HID class, this commit add the support for the Get_Report request.